register client component nested inside server component
